+++ seqspec/seqspec_onlist.py
@@ -3,7 +3,7 @@ from seqspec.Region import project_regions_to_coordinates, itx_read, Onlist
 from seqspec.utils import load_spec, map_read_id_to_regions
 from seqspec.seqspec_find import run_find_by_type, run_find
 import os
-from seqspec.utils import read_list, find_onlist_file
+from seqspec.utils import read_local_list, read_remote_list
 import itertools
 from typing import List
 
@@ -67,8 +67,16 @@ def validate_onlist_args(parser, args):
     m = args.m
     r = args.r
     f = args.f
-    # TODO: if onlist is a link, download. also fix output path
-    # o = args.o
+
+    # the base path is the path to the spec file
+    base_path = os.path.dirname(os.path.abspath(fn))
+
+    # set the save path if it exists
+    save_path = args.o
+    if not save_path:
+        # otherwise the save path is the same path as the spec
+        save_path = base_path
+
     # load spec
     spec = load_spec(fn)
     # if number of barcodes > 1 then we need to join them
@@ -79,68 +87,97 @@ def validate_onlist_args(parser, args):
         for ol in onlists:
             print(f"{ol['region_id']}\t{ol['filename']}\t{ol['location']}\t{ol['md5']}")
         return
-    if args.s == "region":
-        olist = run_onlist_region(spec, m, r, f)
-    elif args.s == "region-type":
-        olist = run_onlist_region_type(spec, m, r, f)
-    elif args.s == "read":
-        olist = run_onlist_read(spec, m, r, f)
-    print(os.path.join(os.path.dirname(os.path.abspath(fn)), olist))
+
+    CMD = {
+        "region": run_onlist_region,
+        "region-type": run_onlist_region_type,
+        "read": run_onlist_read,
+    }
+    if args.s in CMD.keys():
+        onlists = CMD[args.s](spec, m, r)
+    else:
+        raise ValueError(f"Unrecognized spec object {args.s}")
+
+    if len(onlists) == 0:
+        raise ValueError(f"No onlist found for {m}, {args.s}, {r}")
+
+    # for only one onlist we can just return the path
+    # if only one, its remote and we save it to the base path
+    elif len(onlists) == 1:
+        location = onlists[0].location
+        onlist_fn = os.path.basename(onlists[0].filename)
+        if location == "local":
+            onlist_path = os.path.join(base_path, onlist_fn)
+        elif location == "remote":
+            # download the onlist to the base path and return the path
+            onlist_elements = read_remote_list(onlists[0], save_path)
+            onlist_path = os.path.join(
+                write_onlist(onlist_elements, save_path), onlist_fn
+            )
+
+    # anytime we join onlists, we create a new onlist file
+    elif len(onlists) > 1:
+        lsts = []
+        for o in onlists:
+            if o.location == "local":
+                lsts.append(read_local_list(o, base_path))
+            elif o.location == "remote":
+                # base_path is ignored for remote onlists
+                lsts.append(read_remote_list(o, base_path))
+        onlist_elements = join_onlists(onlists, f)
+        onlist_path = os.path.join(
+            write_onlist(onlist_elements, save_path), "onlist_joined.txt"
+        )
+
+    # print the path to the onlist
+    print(onlist_path)
     return
 
 
-def run_onlist_region_type(spec: Assay, modality: str, region_type: str, fmt: str):
-    # for now return the path to the onlist file for the modality/region pair
-
-    # run function
+def run_onlist_region_type(
+    spec: Assay, modality: str, region_type: str
+) -> List[Onlist]:
     regions = run_find_by_type(spec, modality, region_type)
-    onlists = []
+    onlists: List[Onlist] = []
     for r in regions:
-        onlists.append(r.get_onlist())
-    if len(onlists) == 0:
-        raise ValueError(f"No onlist found for region type {region_type}")
-    return join_onlists(onlists, fmt)
+        ol = r.get_onlist()
+        if ol:
+            onlists.append(ol)
+    return onlists
 
 
-def run_onlist_region(spec: Assay, modality: str, region_id: str, fmt: str):
-    # for now return the path to the onlist file for the modality/region pair
-
-    # run function
+def run_onlist_region(spec: Assay, modality: str, region_id: str) -> List[Onlist]:
     regions = run_find(spec, modality, region_id)
-    onlists = []
+    onlists: List[Onlist] = []
     for r in regions:
         onlists.append(r.get_onlist())
     if len(onlists) == 0:
         raise ValueError(f"No onlist found for region {region_id}")
-    return join_onlists(onlists, fmt)
+    return onlists
 
 
-def run_onlist_read(spec: Assay, modality: str, read_id: str, fmt: str):
-    # for now return the path to the onlist file for the modality/region pair
-
-    # run function
+def run_onlist_read(spec: Assay, modality: str, read_id: str) -> List[Onlist]:
     (read, rgns) = map_read_id_to_regions(spec, modality, read_id)
     # convert regions to region coordinates
     rcs = project_regions_to_coordinates(rgns)
     # intersect read with region coordinates
     new_rcs = itx_read(rcs, 0, read.max_len)
 
-    onlists = []
+    onlists: List[Onlist] = []
     for r in new_rcs:
         ol = r.get_onlist()
         if ol:
             onlists.append(ol)
 
-    if len(onlists) == 0:
-        raise ValueError(f"No onlist found for read {read_id}")
-
-    return join_onlists(onlists, fmt)
+    return onlists
 
 
 def run_list_onlists(spec: Assay, modality: str):
     regions = spec.get_libspec(modality).get_onlist_regions()
     olsts = []
     for r in regions:
+        if r.onlist is None:
+            continue
         olsts.append(
             {
                 "region_id": r.region_id,
@@ -162,43 +199,32 @@ def find_list_target_dir(onlists):
     return os.getcwd()
 
 
-def join_onlists(onlists: List[Onlist], fmt: str):
+def join_onlists(onlists: List[List[str]], fmt: str) -> List[str]:
     """Given a list of onlist objects return a file containing the combined list"""
-    if len(onlists) == 0:
-        print("No lists present")
-        return
 
-    # look to see if the barcode file is present.
-    first_location, first_filename = find_onlist_file(onlists[0])
-    if len(onlists) == 1 and first_location == "local":
-        return first_filename
-    else:
-        base_path = find_list_target_dir(onlists)
-        # join the onlists
-        lsts = [read_list(o) for o in onlists]
-        joined_path = os.path.join(base_path, "onlist_joined.txt")
-        formatter_functions = {
-            "product": join_product_onlist,
-            "multi": join_multi_onlist,
-        }
-        formatter = formatter_functions.get(fmt)
-        if formatter is None:
-            raise ValueError(
-                f"Unrecognized format type {fmt}. Expected {', '.join(list(formatter_functions.keys()))}"
-            )
+    # base path should be the path relative to the spec file
+    # join the onlists
+    formatter_functions = {
+        "product": join_product_onlist,
+        "multi": join_multi_onlist,
+    }
+    joined_onlist = list(formatter_functions[fmt](onlists))
 
-        with open(joined_path, "w") as f:
-            for line in formatter(lsts):
-                f.write(line)
-
-        return joined_path
+    return joined_onlist
 
 
-def join_product_onlist(lsts):
+def write_onlist(onlist: List[str], path: str) -> str:
+    with open(path, "w") as f:
+        for line in onlist:
+            f.write(line)
+    return path
+
+
+def join_product_onlist(lsts: List[List[str]]):
     for i in itertools.product(*lsts):
         yield f"{''.join(i)}\n"
 
 
-def join_multi_onlist(lsts):
+def join_multi_onlist(lsts: List[List[str]]):
     for row in itertools.zip_longest(*lsts, fillvalue="-"):
         yield f"{' '.join((str(x) for x in row))}\n"